move deep link logic to Expensify
